MySQL-EX/app/requirements.txt
flask==3.1.3
python-dotenv==1.0.1
mysql-connector-python==9.1.0
